[UPBE-386] notificationDeleteOk only for recovery

+++ src/main/java/it/gov/pagopa/notification/manager/service/NotificationManagerServiceImpl.java
@@ -163,7 +163,6 @@ public class NotificationManagerServiceImpl implements NotificationManagerServic
         }
 
         logNotificationId(notificationId);
-//        notificationSent(notification);
     }
 
     @Override
@@ -241,7 +240,7 @@ public class NotificationManagerServiceImpl implements NotificationManagerServic
 
         logNotificationId(notificationId);
 
-//        notificationSent(notification);
+        notificationSent(notification);
         performanceLog(startTime);
 
         return true;
@@ -446,8 +445,6 @@ public class NotificationManagerServiceImpl implements NotificationManagerServic
 
         logNotificationId(notificationId);
 
-//        notificationSent(notification);
-
         performanceLog(startTime);
     }
 
@@ -508,7 +505,9 @@ public class NotificationManagerServiceImpl implements NotificationManagerServic
     }
 
     private void notificationSent(Notification notification) {
-        notificationManagerRepository.deleteById(notification.getId());
+        if(notification.getId() != null) {
+            notificationManagerRepository.deleteById(notification.getId());
+        }
     }
 
     private void performanceLog(long startTime) {
